Refactor and use conversations api format
